Fix for HP-UX.

+++ builtin.c
@@ -1558,19 +1558,19 @@ mpf1:
 			case MP_INT_WITH_PREC:
 				sprintf(cp, "*.*Z%c", cs1);
 				while ((nc = mpfr_snprintf(obufout, ofre, cpbuf,
-					     (int) fw, (int) prec, zi)) >= ofre)
+					     (int) fw, (int) prec, zi)) >= (int) ofre)
 					chksize(nc)
 				break;
 			case MP_INT_WITHOUT_PREC:
 				sprintf(cp, "*Z%c", cs1);
 				while ((nc = mpfr_snprintf(obufout, ofre, cpbuf,
-					     (int) fw, zi)) >= ofre)
+					     (int) fw, zi)) >= (int) ofre)
 					chksize(nc)
 				break;
 			case MP_FLOAT:
 				sprintf(cp, "*.*R*%c", cs1);
 				while ((nc = mpfr_snprintf(obufout, ofre, cpbuf,
-					     (int) fw, (int) prec, ROUND_MODE, mf)) >= ofre)
+					     (int) fw, (int) prec, ROUND_MODE, mf)) >= (int) ofre)
 					chksize(nc)
 				break;
 #endif
@@ -1579,7 +1579,7 @@ mpf1:
 					sprintf(cp, "*.*%c", cs1);
 					while ((nc = snprintf(obufout, ofre, cpbuf,
 						     (int) fw, (int) prec,
-						     (double) tmpval)) >= ofre)
+						     (double) tmpval)) >= (int) ofre)
 						chksize(nc)
 				} else {
 					// For %a and %A, use the default precision if it
@@ -1587,7 +1587,7 @@ mpf1:
 					sprintf(cp, "*%c", cs1);
 					while ((nc = snprintf(obufout, ofre, cpbuf,
 						     (int) fw,
-						     (double) tmpval)) >= ofre)
+						     (double) tmpval)) >= (int) ofre)
 						chksize(nc)
 				}
 			}
